+++ cmd/camt/convert_test.go
@@ -37,10 +37,6 @@ func (m *mockFullParser) SetLogger(logger logging.Logger) {}
 
 func (m *mockFullParser) SetCategorizer(categorizer models.TransactionCategorizer) {}
 
-func (m *mockFullParser) BatchConvert(ctx context.Context, inputDir, outputDir string) (int, error) {
-	return 0, nil
-}
-
 // Ensure mockFullParser implements parser.FullParser
 var _ parser.FullParser = (*mockFullParser)(nil)